also handle error in the log service schema storage

+++ logservice/logpuller/log_puller.go
@@ -181,7 +181,7 @@ func (p *LogPuller) Subscribe(
 	p.subscriptions.spanProgressMap[subID] = progress
 	p.subscriptions.Unlock()
 
-	p.client.Subscribe(subID, span, uint64(startTs))
+	p.client.Subscribe(subID, span, startTs)
 	return subID
 }
